requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
transformers==4.46.3
torch==2.5.1
pillow==11.0.0
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.10.2
pydantic-settings==2.6.1
